move context provider and host elements into src/core folder to avoid exposing them to the public
