--- libs/element/index.out.ts
@@ -1,1 +1,0 @@
-export * from './element.mixin';
