+++ libs/element/index.ts
@@ -1,4 +1,2 @@
 export * from './element.mixin';
 export * from './lit-element.element';
-export * from './context-provider.element';
-export * from './controller-host.element';
