+++ src/core/context-provider/context-provider.element.test.ts
@@ -1,7 +1,7 @@
 import { expect, fixture, html } from '@open-wc/testing';
 import { customElement } from 'lit/decorators.js';
 import { UmbContextProviderElement } from './context-provider.element';
-import { UmbLitElement } from './lit-element.element';
+import { UmbLitElement } from '@umbraco-cms/element';
 
 @customElement('umb-context-test')
 export class ContextTestElement extends UmbLitElement {
